TaskExecutor/requirements.txt
pyautogui
pillow
python-dotenv
requests
argparse